+++ test/text_anchors_multiline.py
@@ -10,9 +10,9 @@ from imagesmacker.models.draw import TextConfig, TextStyle
 from imagesmacker.models.fields import (
     FieldsConfig,
     FieldsCoords,
+    RelativeContainer,
     RelativeDataFieldFormat,
     RelativeFieldCell,
-    RelativeRow,
     TextFieldAttributes,
 )
 
@@ -66,9 +66,9 @@ for y_anchor in vertical_anchors:
                 style=TextStyle(fill="#fff"),
             ),
         )
-    data_field_fmt_rows.append(RelativeRow(fr=1, cells=y_anchor_list))
+    data_field_fmt_rows.append(RelativeContainer(fr=1, cells=y_anchor_list))
 
-data_field_fmt = RelativeDataFieldFormat(rows=data_field_fmt_rows)
+data_field_fmt = RelativeDataFieldFormat(cells=data_field_fmt_rows)
 
 fields_coords = relative_field_formatting(
     data_field_fmt,
@@ -91,6 +91,7 @@ def draw_boxes() -> None:
 draw_boxes()
 image.save(
     os.path.join(
-        parent_dir_nth_times(__file__, 2), "docs/examples/text-anchors-multiline.png",
+        parent_dir_nth_times(__file__, 2),
+        "docs/examples/text-anchors-multiline.png",
     ),
 )